+++ templates/login.html
@@ -1,4 +1,5 @@
-<html>
+{% extends "navbar.html" %}
+{% block content %}
     <head>
         <title>Page d'accueil</title>
         <link rel="stylesheet" type="text/css" href="../static/MonFichierStyle.css">
@@ -17,4 +18,4 @@
         <div id="cadre">{{ message }}"</div>
         {% endif %}
     </body>
-</html>
+{% endblock %}
