+++ templates/navbar.html
@@ -9,6 +9,7 @@
                 <h1 class="title"> MagicShop</h1>
                 <div class="links">
                     <a href="http://localhost:5000"> Home </a>
+                    <a href="http://localhost:5000/signup"> SignUp </a>
                     <a href="http://localhost:5000/catalog"> Catalog </a>
                 </div>
             </div>
