+++ templates/signup.html
@@ -1,4 +1,5 @@
-<html>
+{% extends "navbar.html" %}
+{% block content %}
     <head>
         <title>Page d'inscription</title>
         <script type="text/javascript" src="../static/MonFichierCode.js"></script>
@@ -17,4 +18,4 @@
         <div>{{ message }}</div>
         {% endif %}
 </body>
-</html>
+{% endblock %}
